requirements.txt
Django==1.11.14
pytz==2018.5
PyYAML==3.13
